% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runScreening.R
\docType{package}
\name{AbstractScreening}
\alias{AbstractScreening}
\alias{AbstractScreening-package}
\title{Shiny App for Abstract Screening}
\description{
Shiny App for Abstract Screening using multiple MLT methods.
}
\author{
Stijn Jaspers
}
