% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runScreening.R
\name{runAbstractScreening}
\alias{runAbstractScreening}
\title{Launch the Shiny App for Abstract Screening}
\usage{
runAbstractScreening()
}
\description{
Shiny App for Abstract Screening using multiple MLT methods.
}
\examples{
\dontrun{#'
## FIRST-TIME INSTALLATION ##
## Note: If updating Shiny App, skip to step 4
# 1. Install R (>= 3.3.0)
# 2. Install Appropriate version of R Tools (https://cran.r-project.org/bin/windows/Rtools/)
# 3. Install devtools R Package
install.packages("devtools")
devtools::install_github("hadley/devtools")
# 4. Install Abstract Screening Package (from the directory with source files)
devtools::install(".../AbstractScreening")
# 5. If you do not have the development version of the caretEnsemble R package, run
devtools::install_github('zachmayer/caretEnsemble')

## LAUNCHING THE APP ##
library(AbstractScreening)
runAbstractScreening() # Note: Only necessary if you need to launch the app again.


# REMARK EWOUD:
# THE ABOVE INSTRUCTIONS ARE IF WE WANT TO DISTRIBUTE THE PACKAGE THROUGH THE SOURCE FILES
# IF THEY WANT NOT INSTALL, BUT JUST A ZIP OF THE BINARIES, WE NEED TO EMPLOY A DIFFERENT
# STRATEGY (.onAttach to install dependencies) However caretEnsemble additional install
# (due to needing developer version), still applies (because same version!).

}
}
\author{
Stijn Jaspers
}
